# Speed up messages split by 10x