Before:
```
make split-PaperMarioMessages  32.03s user 4.69s system 98% cpu 37.138 total
```


After:
```
make split-PaperMarioMessages  3.71s user 0.32s system 98% cpu 4.085 total
```

(holy shit)